Add book adding functionality
Added functions for resetting the form inputs, adding a number property to the
created books, so they're linked to their place in the array. Added a remove
button to created book cards, when clicked it will remove the book from the
library array and page and assign the other cards new numbers to match their
location in the array. Added a toggle read button on the card, that toggles
its status in the library array. Added styling to the page.